+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/state/RepairStateImpl.java
@@ -208,11 +208,9 @@ public class RepairStateImpl implements RepairState
         long initialDelayInMs = myRepairConfiguration.getInitialDelayInMs();
         long assumedRepairedAt = System.currentTimeMillis() - runIntervalInMs + initialDelayInMs;
 
-        if (LOG.isInfoEnabled())
-        {
-            LOG.info("Assuming the table {} is new. Next repair will occur at {}",
-                    myTableReference, MY_DATE_FORMAT.get().format(new Date(assumedRepairedAt + runIntervalInMs)));
-        }
+        LOG.info("Assuming the table {} is new. Next repair will occur at {}.",
+                myTableReference,
+                MY_DATE_FORMAT.get().format(new Date(assumedRepairedAt + runIntervalInMs)));
 
         return assumedRepairedAt;
     }