+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 ## Version 5.0.4
 
+* ecChronos will break if repair interval is shorter than the initial delay - Issue #667
+
 ## Version 5.0.3
 
 * Spring Framework URL Parsing with Host Validation is vulnerable - Issue #661